+++ lib/metrics/statsd_legacy.go
@@ -56,9 +56,10 @@ func (s *StatsdLegacyStat) Set(value int64) error {
 // StatsdLegacy is a stats object with capability to hold internal stats as a JSON
 // endpoint.
 type StatsdLegacy struct {
-	config Config
-	s      statsd.Statsd
-	log    log.Modular
+	config      Config
+	s           statsd.Statsd
+	log         log.Modular
+	pathMapping *pathMapping
 }
 
 // NewStatsdLegacy creates and returns a new StatsdLegacy object.
@@ -73,6 +74,10 @@ func NewStatsdLegacy(config Config, opts ...func(Type)) (Type, error) {
 	}
 	for _, opt := range opts {
 		opt(s)
+	}
+
+	if s.pathMapping, err = newPathMapping(config.Statsd.PathMapping, s.log); err != nil {
+		return nil, fmt.Errorf("failed to init path mapping: %v", err)
 	}
 
 	if config.Statsd.Network == "tcp" {
@@ -123,6 +128,9 @@ func NewStatsdLegacy(config Config, opts ...func(Type)) (Type, error) {
 
 // GetCounter returns a stat counter object for a path.
 func (h *StatsdLegacy) GetCounter(path string) StatCounter {
+	if path = h.pathMapping.mapPath(path); len(path) == 0 {
+		return DudStat{}
+	}
 	return &StatsdLegacyStat{
 		path: path,
 		s:    h.s,
@@ -132,7 +140,11 @@ func (h *StatsdLegacy) GetCounter(path string) StatCounter {
 // GetCounterVec returns a stat counter object for a path with the labels
 // discarded.
 func (h *StatsdLegacy) GetCounterVec(path string, n []string) StatCounterVec {
+	path = h.pathMapping.mapPath(path)
 	return fakeCounterVec(func([]string) StatCounter {
+		if len(path) == 0 {
+			return DudStat{}
+		}
 		return &StatsdLegacyStat{
 			path: path,
 			s:    h.s,
@@ -142,6 +154,9 @@ func (h *StatsdLegacy) GetCounterVec(path string, n []string) StatCounterVec {
 
 // GetTimer returns a stat timer object for a path.
 func (h *StatsdLegacy) GetTimer(path string) StatTimer {
+	if path = h.pathMapping.mapPath(path); len(path) == 0 {
+		return DudStat{}
+	}
 	return &StatsdLegacyStat{
 		path: path,
 		s:    h.s,
@@ -151,7 +166,11 @@ func (h *StatsdLegacy) GetTimer(path string) StatTimer {
 // GetTimerVec returns a stat timer object for a path with the labels
 // discarded.
 func (h *StatsdLegacy) GetTimerVec(path string, n []string) StatTimerVec {
+	path = h.pathMapping.mapPath(path)
 	return fakeTimerVec(func([]string) StatTimer {
+		if len(path) == 0 {
+			return DudStat{}
+		}
 		return &StatsdLegacyStat{
 			path: path,
 			s:    h.s,
@@ -161,6 +180,9 @@ func (h *StatsdLegacy) GetTimerVec(path string, n []string) StatTimerVec {
 
 // GetGauge returns a stat gauge object for a path.
 func (h *StatsdLegacy) GetGauge(path string) StatGauge {
+	if path = h.pathMapping.mapPath(path); len(path) == 0 {
+		return DudStat{}
+	}
 	return &StatsdLegacyStat{
 		path: path,
 		s:    h.s,
@@ -170,7 +192,11 @@ func (h *StatsdLegacy) GetGauge(path string) StatGauge {
 // GetGaugeVec returns a stat timer object for a path with the labels
 // discarded.
 func (h *StatsdLegacy) GetGaugeVec(path string, n []string) StatGaugeVec {
+	path = h.pathMapping.mapPath(path)
 	return fakeGaugeVec(func([]string) StatGauge {
+		if len(path) == 0 {
+			return DudStat{}
+		}
 		return &StatsdLegacyStat{
 			path: path,
 			s:    h.s,