+++ lib/metrics/whitelist.go
@@ -17,6 +17,7 @@ import (
 func init() {
 	Constructors[TypeWhiteList] = TypeSpec{
 		constructor: NewWhitelist,
+		Deprecated:  true,
 		Summary: `
 Whitelist metric paths within Benthos so that only matching metric paths are
 aggregated by a child metric target.`,